+++ src/main/java/com/academia/acadDependences/controller/ProfessorController.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.academia.acadDependences.model.Aluno;
 import com.academia.acadDependences.model.Professor;
 import com.academia.acadDependences.model.Professor;
 import com.academia.acadDependences.repository.ProfessorRepository;
@@ -26,10 +29,10 @@ public class ProfessorController {
 	@Autowired
 	private ProfessorRepository professorRepository;
 	private List<Professor> Professors = new ArrayList<Professor>();
-	
+
 	ProfessorController(ProfessorRepository professorRepository) {
-        this.professorRepository = professorRepository;
-    }
+		this.professorRepository = professorRepository;
+	}
 
 	@GetMapping("/MostrarLista")
 	public List<Professor> viewProfessors() {
@@ -38,36 +41,42 @@ public class ProfessorController {
 
 	@GetMapping("/procurarProfessor/{id}")
 	public Optional<Professor> viewProfessor(@PathVariable("id") int id) {
-		return  this.professorRepository.findById(id);
-		}
-
+		return this.professorRepository.findById(id);
+	}
 
 	@PutMapping("/atualizar/{id}")
 	public void atualizarProfessor(@PathVariable("id") int id, @RequestBody Professor newProfessor) {
 		Optional<Professor> ProfessorFind = this.professorRepository.findById(id);
 		if (ProfessorFind.isPresent()) {
-			
-			
+
 			Professor professor = ProfessorFind.get();
 			professor.setName(newProfessor.getName());
 			professor.setCpf(newProfessor.getCpf());
-            professor.setIdade(newProfessor.getIdade());
-            professor.setSalario(newProfessor.getSalario());
-            professorRepository.save(professor);
+			professor.setIdade(newProfessor.getIdade());
+			professor.setSalario(newProfessor.getSalario());
+			professorRepository.save(professor);
+			System.out.println("Professor atualizado com sucesso!");
+		} else {
+			System.out.println("Professor nao encontrado!");
 		}
-		
+
 	}
 
 	@PostMapping("/create")
-	void criarProfessor(@RequestBody Professor professor) {
+	public void criarProfessor(@Valid @RequestBody Professor professor) {
 		professorRepository.save(professor);
 
 	}
 
 	@DeleteMapping("/delete/{id}")
-	void delete(@PathVariable("id") int id) {
-	professorRepository.deleteById(id);
-	}
-	}
+	public String deleteprofessor(@PathVariable("id") int id) {
+		Optional<Professor> professorFind = this.professorRepository.findById(id);
+		if (professorFind.isPresent()) {
+			professorRepository.deleteById(id);
+			return ("Professor Deletado com sucesso");
+		} else {
+			return ("Professor Nao Encontrado");
+		}
 
-
+	}
+}
